refactor(financeiro): extrai leaf components de FinanceiroDashboard (1297→819)
Quebra do god-component (§10). Move helpers puros e os sub-componentes de
apresentação já inline para src/components/financeiro/dashboard/:
- format.ts (fmt, fmtCompact, fmtDate, statusColor, getWeekLabel, formatCnpj)
- KpiCard, AgingCard, FluxoCaixaTab, DRETab, DREComparativo

Movimentação 1:1 verbatim (inclui os `any` pré-existentes, sem mexer em strict).
Comportamento preservado; imports mortos removidos (FileText, PieChart,
Skeleton, getWeekLabel). Página 1297 → 819 (<1000). tsc limpo, 423/423 testes.

A extração dos 3 TabsContent grandes do componente principal fica como
follow-up (maior risco de prop-drilling em tela financeira).
