+++ src/pages/FinanceiroDashboard.tsx
@@ -5,7 +5,6 @@ import { Button } from '@/components/ui/button';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
-import { Skeleton } from '@/components/ui/skeleton';
 import { useFinanceiro, type FinanceiroView } from '@/hooks/useFinanceiro';
 import { useFinanceiroRegime } from '@/hooks/useFinanceiroRegime';
 import { COMPANIES, ALL_COMPANIES, type Company } from '@/contexts/CompanyContext';
@@ -15,40 +14,21 @@ import {
 import {
   Loader2, RefreshCw, DollarSign, TrendingUp, TrendingDown,
   AlertTriangle, Wallet, ArrowDownCircle, ArrowUpCircle,
-  Building2, BarChart3, PieChart, Calendar, FileText,
+  Building2, BarChart3, Calendar,
   Clock, Download, History
 } from 'lucide-react';
 import { AuditTrailDrawer } from '@/components/financeiro/AuditTrailDrawer';
 import { usePeriodLockHandler } from '@/components/financeiro/PeriodLockGuard';
 import { generateAlerts } from '@/utils/financeiroAlerts';
 import { RegimeToggle } from '@/components/financeiro/RegimeToggle';
-
-// ═══════════════ FORMATTERS ═══════════════
-
-const fmt = (v: number) =>
-  v.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' });
-
-const fmtCompact = (v: number) => {
-  if (Math.abs(v) >= 1_000_000) return `R$ ${(v / 1_000_000).toFixed(1)}M`;
-  if (Math.abs(v) >= 1_000) return `R$ ${(v / 1_000).toFixed(1)}k`;
-  return fmt(v);
-};
-
-const fmtDate = (d: string | null) => {
-  if (!d) return '—';
-  const date = new Date(d + 'T00:00:00');
-  return date.toLocaleDateString('pt-BR');
-};
-
-const statusColor = (s: string) => {
-  switch (s) {
-    case 'PAGO': case 'RECEBIDO': case 'LIQUIDADO': return 'bg-status-success-bg text-status-success';
-    case 'VENCIDO': return 'bg-status-error-bg text-status-error';
-    case 'PARCIAL': return 'bg-status-warning-bg text-status-warning';
-    case 'CANCELADO': return 'bg-gray-100 text-gray-500';
-    default: return 'bg-status-info-bg text-status-info';
-  }
-};
+import {
+  fmt, fmtCompact, fmtDate, statusColor, formatCnpj,
+} from '@/components/financeiro/dashboard/format';
+import { KpiCard } from '@/components/financeiro/dashboard/KpiCard';
+import { AgingCard } from '@/components/financeiro/dashboard/AgingCard';
+import { FluxoCaixaTab } from '@/components/financeiro/dashboard/FluxoCaixaTab';
+import { DRETab } from '@/components/financeiro/dashboard/DRETab';
+import { DREComparativo } from '@/components/financeiro/dashboard/DREComparativo';
 
 const today = new Date();
 const sixMonthsAgo = new Date(today.getFullYear(), today.getMonth() - 6, 1);
@@ -835,463 +815,5 @@ const FinanceiroDashboard = () => {
     </div>
   );
 };
-
-// ═══════════════ SUB-COMPONENTS ═══════════════
-
-function KpiCard({ title, value, icon: Icon, color, bgColor, subtitle, subtitleColor }: {
-  title: string;
-  value: number;
-  icon: any;
-  color: string;
-  bgColor: string;
-  subtitle?: string;
-  subtitleColor?: string;
-}) {
-  return (
-    <Card>
-      <CardContent className="p-4">
-        <div className="flex items-start justify-between">
-          <div>
-            <p className="text-xs text-muted-foreground font-medium">{title}</p>
-            <p className={`text-lg font-bold mt-1 ${color}`}>{fmtCompact(value)}</p>
-            {subtitle && (
-              <p className={`text-xs mt-1 ${subtitleColor || 'text-muted-foreground'}`}>{subtitle}</p>
-            )}
-          </div>
-          <div className={`p-2 rounded-lg ${bgColor}`}>
-            <Icon className={`w-4 h-4 ${color}`} />
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-function AgingCard({ title, data }: { title: string; data: any; type: 'receber' | 'pagar' }) {
-  if (!data) return (
-    <Card>
-      <CardHeader className="pb-3">
-        <CardTitle className="text-base">{title}</CardTitle>
-      </CardHeader>
-      <CardContent><Skeleton className="h-40" /></CardContent>
-    </Card>
-  );
-
-  const total =
-    data.a_vencer_valor +
-    data.vencido_1_30_valor +
-    data.vencido_31_60_valor +
-    data.vencido_61_90_valor +
-    data.vencido_90_plus_valor;
-
-  const bars = [
-    { label: 'A vencer', value: data.a_vencer_valor, qtd: data.a_vencer_qtd, color: 'bg-status-info' },
-    { label: '1-30 dias', value: data.vencido_1_30_valor, qtd: data.vencido_1_30_qtd, color: 'bg-status-warning' },
-    { label: '31-60 dias', value: data.vencido_31_60_valor, qtd: data.vencido_31_60_qtd, color: 'bg-status-warning-bold' },
-    { label: '61-90 dias', value: data.vencido_61_90_valor, qtd: data.vencido_61_90_qtd, color: 'bg-status-error' },
-    { label: '+90 dias', value: data.vencido_90_plus_valor, qtd: data.vencido_90_plus_qtd, color: 'bg-status-error' },
-  ];
-
-  return (
-    <Card>
-      <CardHeader className="pb-3">
-        <CardTitle className="text-base">{title}</CardTitle>
-        <p className="text-xs text-muted-foreground">Total: {fmt(total)}</p>
-      </CardHeader>
-      <CardContent className="space-y-3">
-        {bars.map(b => (
-          <div key={b.label} className="space-y-1">
-            <div className="flex justify-between text-xs">
-              <span className="text-muted-foreground">{b.label} ({b.qtd})</span>
-              <span className="font-medium">{fmtCompact(b.value)}</span>
-            </div>
-            <div className="h-2 rounded-full bg-muted overflow-hidden">
-              <div
-                className={`h-full rounded-full ${b.color} transition-all`}
-                style={{ width: total > 0 ? `${Math.max((b.value / total) * 100, 1)}%` : '0%' }}
-              />
-            </div>
-          </div>
-        ))}
-      </CardContent>
-    </Card>
-  );
-}
-
-function FluxoCaixaTab({ data, loading, saldoCC }: { data: any[]; loading: boolean; saldoCC?: number }) {
-  if (loading) return <Skeleton className="h-60" />;
-  if (!data || data.length === 0) {
-    return (
-      <Card>
-        <CardContent className="py-12 text-center text-muted-foreground">
-          <BarChart3 className="w-10 h-10 mx-auto mb-3 opacity-40" />
-          Nenhum dado de fluxo de caixa. Sincronize os dados primeiro.
-        </CardContent>
-      </Card>
-    );
-  }
-
-  const todayStr = new Date().toISOString().slice(0, 10);
-
-  // Separar passado (realizado) e futuro (previsto)
-  const totalEntradasRealizadas = data.reduce((s, d) => s + (d.entradas_realizadas || 0), 0);
-  const totalSaidasRealizadas = data.reduce((s, d) => s + (d.saidas_realizadas || 0), 0);
-  const totalEntradasPrevistas = data
-    .filter(d => d.data >= todayStr)
-    .reduce((s, d) => s + (d.entradas_previstas || 0), 0);
-  const totalSaidasPrevistas = data
-    .filter(d => d.data >= todayStr)
-    .reduce((s, d) => s + (d.saidas_previstas || 0), 0);
-
-  // Agrupar por semana para simplificar visualização
-  const weeks: { label: string; entradas: number; saidas: number; saldo: number; acumulado: number }[] = [];
-  let weekEntradas = 0, weekSaidas = 0;
-  let currentWeek = '';
-  let acumulado = saldoCC || 0;
-
-  for (const day of data) {
-    const d = new Date(day.data + 'T00:00:00');
-    const weekNum = getWeekLabel(d);
-    if (weekNum !== currentWeek && currentWeek !== '') {
-      acumulado += weekEntradas - weekSaidas;
-      weeks.push({ label: currentWeek, entradas: weekEntradas, saidas: weekSaidas, saldo: weekEntradas - weekSaidas, acumulado });
-      weekEntradas = 0;
-      weekSaidas = 0;
-    }
-    currentWeek = weekNum;
-    const isPast = day.data < todayStr;
-    weekEntradas += isPast ? (day.entradas_realizadas || 0) : (day.entradas_previstas || 0);
-    weekSaidas += isPast ? (day.saidas_realizadas || 0) : (day.saidas_previstas || 0);
-  }
-  if (currentWeek) {
-    acumulado += weekEntradas - weekSaidas;
-    weeks.push({ label: currentWeek, entradas: weekEntradas, saidas: weekSaidas, saldo: weekEntradas - weekSaidas, acumulado });
-  }
-
-  const maxVal = Math.max(...weeks.map(w => Math.max(w.entradas, w.saidas)), 1);
-
-  return (
-    <div className="space-y-4">
-      {/* Summary KPIs */}
-      <div className="grid grid-cols-2 lg:grid-cols-5 gap-3">
-        {saldoCC != null && (
-          <div className="p-3 rounded-lg bg-status-info-bg text-center">
-            <p className="text-xs text-muted-foreground">Saldo CC Atual</p>
-            <p className="text-sm font-bold text-status-info">{fmtCompact(saldoCC)}</p>
-          </div>
-        )}
-        <div className="p-3 rounded-lg bg-status-success-bg text-center">
-          <p className="text-xs text-muted-foreground">Recebido</p>
-          <p className="text-sm font-bold text-status-success">{fmtCompact(totalEntradasRealizadas)}</p>
-        </div>
-        <div className="p-3 rounded-lg bg-status-error-bg text-center">
-          <p className="text-xs text-muted-foreground">Pago</p>
-          <p className="text-sm font-bold text-status-error">{fmtCompact(totalSaidasRealizadas)}</p>
-        </div>
-        <div className="p-3 rounded-lg bg-status-success-bg/50 text-center">
-          <p className="text-xs text-muted-foreground">Previsto Entrar</p>
-          <p className="text-sm font-bold text-status-success">{fmtCompact(totalEntradasPrevistas)}</p>
-        </div>
-        <div className="p-3 rounded-lg bg-status-error-bg/50 text-center">
-          <p className="text-xs text-muted-foreground">Previsto Sair</p>
-          <p className="text-sm font-bold text-status-error">{fmtCompact(totalSaidasPrevistas)}</p>
-        </div>
-      </div>
-
-      {/* Weekly chart */}
-      <Card>
-        <CardHeader className="pb-3">
-          <CardTitle className="text-base">Fluxo de Caixa Semanal</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <div className="space-y-2">
-            {weeks.slice(-12).map((w, i) => (
-              <div key={i} className="grid grid-cols-[80px_1fr_80px_80px] items-center gap-2 text-sm">
-                <span className="text-xs text-muted-foreground truncate">{w.label}</span>
-                <div className="relative h-6">
-                  <div
-                    className="absolute top-0 h-3 rounded bg-status-success/70"
-                    style={{ width: `${(w.entradas / maxVal) * 100}%` }}
-                  />
-                  <div
-                    className="absolute bottom-0 h-3 rounded bg-status-error/70"
-                    style={{ width: `${(w.saidas / maxVal) * 100}%` }}
-                  />
-                </div>
-                <span className={`text-right text-xs font-bold ${w.saldo >= 0 ? 'text-status-success' : 'text-status-error'}`}>
-                  {fmtCompact(w.saldo)}
-                </span>
-                <span className={`text-right text-[10px] ${w.acumulado >= 0 ? 'text-status-info' : 'text-status-error'}`}>
-                  {fmtCompact(w.acumulado)}
-                </span>
-              </div>
-            ))}
-          </div>
-          <div className="flex gap-4 mt-4 justify-center text-xs text-muted-foreground">
-            <span className="flex items-center gap-1"><span className="w-3 h-3 rounded bg-status-success/70" /> Entradas</span>
-            <span className="flex items-center gap-1"><span className="w-3 h-3 rounded bg-status-error/70" /> Saídas</span>
-            <span>Saldo semanal</span>
-            <span className="text-status-info">Acumulado</span>
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  );
-}
-
-function DRETab({ data, view, ano }: { data: any[]; view: FinanceiroView; ano: number }) {
-  if (!data || data.length === 0) {
-    return (
-      <Card>
-        <CardContent className="py-12 text-center text-muted-foreground">
-          <FileText className="w-10 h-10 mx-auto mb-3 opacity-40" />
-          Nenhum DRE calculado para {ano}. Clique em "Recalcular" para gerar.
-        </CardContent>
-      </Card>
-    );
-  }
-
-  // Data already comes consolidated from hook (dreConsolidado) — no re-consolidation needed
-  const rows = [...data].sort((a: any, b: any) => a.mes - b.mes);
-
-  // Ponto 5: check for unmapped categories
-  const unmappedCats = rows.flatMap((r: any) =>
-    r.detalhamento?.categorias_nao_mapeadas || []
-  );
-  const uniqueUnmapped = [...new Set(unmappedCats)];
-
-  const meses = ['Jan', 'Fev', 'Mar', 'Abr', 'Mai', 'Jun', 'Jul', 'Ago', 'Set', 'Out', 'Nov', 'Dez'];
-
-  const dreLines = [
-    { label: 'Receita Bruta', field: 'receita_bruta', bold: true, color: '' },
-    { label: '(-) Deduções', field: 'deducoes', bold: false, color: 'text-status-error' },
-    { label: '= Receita Líquida', field: 'receita_liquida', bold: true, color: '' },
-    { label: '(-) CMV', field: 'cmv', bold: false, color: 'text-status-error' },
-    { label: '= Lucro Bruto', field: 'lucro_bruto', bold: true, color: 'text-status-success' },
-    { label: '(-) Desp. Operacionais', field: 'despesas_operacionais', bold: false, color: 'text-status-error' },
-    { label: '(-) Desp. Administrativas', field: 'despesas_administrativas', bold: false, color: 'text-status-error' },
-    { label: '(-) Desp. Comerciais', field: 'despesas_comerciais', bold: false, color: 'text-status-error' },
-    { label: '(-) Desp. Financeiras', field: 'despesas_financeiras', bold: false, color: 'text-status-error' },
-    { label: '(+) Rec. Financeiras', field: 'receitas_financeiras', bold: false, color: 'text-status-success' },
-    { label: '= Resultado Operacional', field: 'resultado_operacional', bold: true, color: '' },
-    { label: '(-) Impostos', field: 'impostos', bold: false, color: 'text-status-error' },
-    { label: '= RESULTADO LÍQUIDO', field: 'resultado_liquido', bold: true, color: '' },
-  ];
-
-  return (
-    <div className="space-y-3">
-      {/* Ponto 5: unmapped warning */}
-      {uniqueUnmapped.length > 0 && (
-        <div className="flex items-start gap-2 p-3 rounded-lg bg-status-warning-bg border border-status-warning/30">
-          <AlertTriangle className="w-4 h-4 text-status-warning mt-0.5 shrink-0" />
-          <div>
-            <p className="text-sm font-medium text-status-warning-fg">
-              {uniqueUnmapped.length} categoria(s) classificadas por heurística
-            </p>
-            <p className="text-xs text-status-warning mt-1">
-              Estas categorias não têm mapeamento explícito — os valores podem estar em linhas incorretas.
-              Configure em <span className="font-medium">Mapeamento DRE</span>.
-            </p>
-            <p className="text-xs text-status-warning mt-1 font-mono">
-              {uniqueUnmapped.slice(0, 8).join(', ')}{uniqueUnmapped.length > 8 ? ` (+${uniqueUnmapped.length - 8})` : ''}
-            </p>
-          </div>
-        </div>
-      )}
-
-      <Card>
-        <CardHeader className="pb-3">
-          <CardTitle className="text-base">
-            DRE Regime de Caixa — {ano}
-            {view === 'all' && <Badge variant="secondary" className="ml-2">Consolidado</Badge>}
-            <Badge variant="outline" className="ml-2 text-[10px]">Regime de Caixa</Badge>
-          </CardTitle>
-        </CardHeader>
-      <CardContent className="p-0">
-        <div className="overflow-x-auto">
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead className="sticky left-0 bg-background min-w-[180px]">Linha</TableHead>
-                {rows.map((r: any) => (
-                  <TableHead key={r.mes} className="text-right min-w-[100px]">
-                    {meses[r.mes - 1]}
-                  </TableHead>
-                ))}
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {dreLines.map(line => (
-                <TableRow key={line.field} className={line.bold ? 'bg-muted/30' : ''}>
-                  <TableCell className={`sticky left-0 bg-background text-sm ${line.bold ? 'font-bold' : ''}`}>
-                    {line.label}
-                  </TableCell>
-                  {rows.map((r: any) => {
-                    const val = r[line.field] || 0;
-                    const colorClass = line.color || (line.bold && line.field.includes('resultado')
-                      ? (val >= 0 ? 'text-status-success' : 'text-status-error')
-                      : '');
-                    return (
-                      <TableCell key={r.mes} className={`text-right text-sm ${line.bold ? 'font-bold' : ''} ${colorClass}`}>
-                        {fmtCompact(val)}
-                      </TableCell>
-                    );
-                  })}
-                </TableRow>
-              ))}
-              {/* Margem bruta % */}
-              <TableRow className="border-t-2">
-                <TableCell className="sticky left-0 bg-background text-sm font-medium text-muted-foreground">
-                  Margem Bruta %
-                </TableCell>
-                {rows.map((r: any) => {
-                  const pct = r.receita_liquida > 0 ? (r.lucro_bruto / r.receita_liquida) * 100 : 0;
-                  return (
-                    <TableCell key={r.mes} className="text-right text-sm text-muted-foreground">
-                      {pct.toFixed(1)}%
-                    </TableCell>
-                  );
-                })}
-              </TableRow>
-              {/* Margem líquida % */}
-              <TableRow>
-                <TableCell className="sticky left-0 bg-background text-sm font-medium text-muted-foreground">
-                  Margem Líquida %
-                </TableCell>
-                {rows.map((r: any) => {
-                  const pct = r.receita_liquida > 0 ? (r.resultado_liquido / r.receita_liquida) * 100 : 0;
-                  return (
-                    <TableCell key={r.mes} className={`text-right text-sm font-medium ${pct >= 0 ? 'text-status-success' : 'text-status-error'}`}>
-                      {pct.toFixed(1)}%
-                    </TableCell>
-                  );
-                })}
-              </TableRow>
-            </TableBody>
-          </Table>
-        </div>
-      </CardContent>
-    </Card>
-    </div>
-  );
-}
-
-// ═══════════════ HELPERS ═══════════════
-
-function DREComparativo({ data, ano }: { data: Record<string, any[]>; ano: number }) {
-  const companies = Object.keys(data);
-  if (companies.length < 2) return null;
-
-  // Calculate annual totals per company
-  const annualTotals = companies.map(co => {
-    const rows = data[co] || [];
-    const total = rows.reduce(
-      (acc, r) => ({
-        receita_liquida: acc.receita_liquida + (r.receita_liquida || 0),
-        lucro_bruto: acc.lucro_bruto + (r.lucro_bruto || 0),
-        resultado_operacional: acc.resultado_operacional + (r.resultado_operacional || 0),
-        resultado_liquido: acc.resultado_liquido + (r.resultado_liquido || 0),
-        impostos: acc.impostos + (r.impostos || 0),
-      }),
-      { receita_liquida: 0, lucro_bruto: 0, resultado_operacional: 0, resultado_liquido: 0, impostos: 0 }
-    );
-    const margemBruta = total.receita_liquida > 0 ? (total.lucro_bruto / total.receita_liquida) * 100 : 0;
-    const margemLiquida = total.receita_liquida > 0 ? (total.resultado_liquido / total.receita_liquida) * 100 : 0;
-    return { company: co, ...total, margemBruta, margemLiquida };
-  });
-
-  const lines: { label: string; field: string; format: 'currency' | 'pct' }[] = [
-    { label: 'Receita Líquida', field: 'receita_liquida', format: 'currency' },
-    { label: 'Lucro Bruto', field: 'lucro_bruto', format: 'currency' },
-    { label: 'Margem Bruta', field: 'margemBruta', format: 'pct' },
-    { label: 'Resultado Operacional', field: 'resultado_operacional', format: 'currency' },
-    { label: 'Impostos', field: 'impostos', format: 'currency' },
-    { label: 'Resultado Líquido', field: 'resultado_liquido', format: 'currency' },
-    { label: 'Margem Líquida', field: 'margemLiquida', format: 'pct' },
-  ];
-
-  return (
-    <Card>
-      <CardHeader className="pb-3">
-        <CardTitle className="text-base flex items-center gap-2">
-          <PieChart className="w-4 h-4" />
-          Comparativo por Empresa — {ano}
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="p-0">
-        <div className="overflow-x-auto">
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead className="sticky left-0 bg-background min-w-[180px]">Indicador</TableHead>
-                {annualTotals.map(t => (
-                  <TableHead key={t.company} className="text-right min-w-[120px]">
-                    <div className="flex flex-col items-end">
-                      <span>{COMPANIES[t.company as Company]?.shortName || t.company}</span>
-                      <Badge variant="outline" className="text-[10px] mt-0.5">
-                        {COMPANIES[t.company as Company]?.regime === 'simples' ? 'SN' : 'LP'}
-                      </Badge>
-                    </div>
-                  </TableHead>
-                ))}
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {lines.map(line => (
-                <TableRow key={line.field}>
-                  <TableCell className="sticky left-0 bg-background text-sm font-medium">
-                    {line.label}
-                  </TableCell>
-                  {annualTotals.map(t => {
-                    const val = (t as any)[line.field] || 0;
-                    const isResult = line.field.includes('resultado') || line.field === 'margemLiquida';
-                    const colorClass = isResult
-                      ? val >= 0 ? 'text-status-success' : 'text-status-error'
-                      : '';
-                    return (
-                      <TableCell key={t.company} className={`text-right text-sm font-medium ${colorClass}`}>
-                        {line.format === 'pct' ? `${val.toFixed(1)}%` : fmtCompact(val)}
-                      </TableCell>
-                    );
-                  })}
-                </TableRow>
-              ))}
-              {/* Participação na receita total */}
-              <TableRow className="border-t-2">
-                <TableCell className="sticky left-0 bg-background text-sm font-medium text-muted-foreground">
-                  % da Receita Total
-                </TableCell>
-                {(() => {
-                  const totalReceita = annualTotals.reduce((s, t) => s + t.receita_liquida, 0);
-                  return annualTotals.map(t => (
-                    <TableCell key={t.company} className="text-right text-sm text-muted-foreground">
-                      {totalReceita > 0 ? `${((t.receita_liquida / totalReceita) * 100).toFixed(1)}%` : '—'}
-                    </TableCell>
-                  ));
-                })()}
-              </TableRow>
-            </TableBody>
-          </Table>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-function getWeekLabel(d: Date): string {
-  const start = new Date(d);
-  start.setDate(start.getDate() - start.getDay());
-  return `${String(start.getDate()).padStart(2, '0')}/${String(start.getMonth() + 1).padStart(2, '0')}`;
-}
-
-function formatCnpj(cnpj: string): string {
-  if (cnpj.length === 14) {
-    return cnpj.replace(/^(\d{2})(\d{3})(\d{3})(\d{4})(\d{2})$/, '$1.$2.$3/$4-$5');
-  }
-  if (cnpj.length === 11) {
-    return cnpj.replace(/^(\d{3})(\d{3})(\d{3})(\d{2})$/, '$1.$2.$3-$4');
-  }
-  return cnpj;
-}
 
 export default FinanceiroDashboard;
